pdf-table-parser-main/requirements.txt
camelot-py==0.9.0
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.53.0
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
opencv-python==4.10.0.84
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.1
pillow==10.3.0
pycparser==2.22
pyparsing==3.1.2
PyPDF2==2.10.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
